07-going-to-production/pyproject.toml
[tool.poetry]
name = "smarttesting-python"
version = "0.0.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.24.0"
factory_boy = "^3.0.1"
injector = "^0.18.3"
pydantic = "^1.6.1"
fastapi = "^0.61.1"
uvicorn = "^0.11.8"
tavern = {extras = ["pytest"], version = "^1.6.0"}
wiremock = "^2.1.2"
sqlalchemy = "^1.3.19"
docker = "^4.3.1"
psycopg2-binary = "^2.8.6"
celery = {extras = ["rabbitmq"], version = "^4.4.7"}
marshmallow_dataclass = "^8.0.0"
marshmallow_enum = "^1.5.1"
kombu = "4.6.11"
flask = "^1.1.2"
Flask-Injector = "^0.12.3"
flask-expects-json = "^1.5.0"
pytest-docker-compose = "^3.2.0"
pymongo = "^3.11.0"
typing_inspect = "^0.6.0"
selenium = "^3.141.0"
selene = {version = "^2.0.0-alpha.33", allow-prereleases = true}
pylint-forbidden-imports = "^1.0.0"
pact-python = "^1.2.6"
pytest-benchmark = "^3.2.3"
locust = "^1.2.3"
pybuilder = "^0.12.9"
invoke = "^1.4.1"
pytest-rerunfailures = "^9.1"
polling = "^0.3.1"
pytest-cov = "^2.10.1"
mutmut = "^2.1.0"
expects = "^0.9.0"
responses = "^0.12.0"
vcrpy = "^4.1.0"
pytest-randomly = "^3.4.1"
assertpy = "^1.1"
Flask-Admin = "^1.5.7"
Flask-Zipkin = "^0.0.4"
prometheus-client = "^0.8.0"
prometheus-flask-exporter = "^0.18.1"

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"
black = "^20.8b1"
isort = "^5.5.1"
mypy = "^0.782"
pylint = "^2.6.0"
flake8 = "^3.8.3"

[tool.isort]
line_length = 88
multi_line_output = 3
include_trailing_comma = true

[tool.pytest.ini_options]
addopts = "-p no:pytest-randomly"
python_files = "test_*.py *_test.py *_tests.py"
markers = "uses_docker wiremock"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
